+++ PetTracker.lua
@@ -46,7 +46,7 @@ end
 --[[ Events ]]--
 
 function Addon:PLAYER_ENTERING_WORLD()
-	PetTracker_Sets = PetTracker_Sets or {}
+	PetTracker_Sets = PetTracker_Sets or {RivalHistory = {}}
 	PetTracker_State = PetTracker_State or {}
 
 	self.Sets = PetTracker_Sets